+++ src/pages/Login.jsx
@@ -1,0 +1,14 @@
+import { CardBody } from '@chakra-ui/react';
+import * as React from 'react';
+
+import Sign from '../component/Sign.jsx';
+
+function Login() {
+  return (
+    <Sign>
+      <CardBody>login</CardBody>
+    </Sign>
+  );
+}
+
+export default Login;
